Created Calculator.js/Moved it in components.
