+++ src/App.test.js
@@ -1,6 +1,6 @@
 /* eslint linebreak-style: 0 */
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from './components/Calculator';
 
 test('renders learn react link', () => {
   render(<App />);
